+++ CHANGES.rst
@@ -22,6 +22,8 @@ Change History
 - Removed solr 3 support
   [do3cc]
 
+- Put log4j properties file into the right location
+  [do3cc]
 
 6.0.0b3 (2015-04-22)
 ====================